[ROCm/rocwmma commit: fde8d3c408be04e366e6d55d1117520c339e11bd]

+++ test/gemm/gemm_PGR1_LB2_MP0_MB_CP/test/ad_hoc_test.cpp
@@ -40,7 +40,7 @@ namespace rocwmma
         // Types: ALL + double
         // Block Sizes: 16 x 16 x BlockK
         // Layouts: NT
-        using Types      = std::tuple<std::tuple<float32_t, float32_t, float32_t>>;
+        using Types      = std::tuple<std::tuple<float16_t, float32_t, float32_t>>;
         using BlockSizes = std::tuple<std::tuple<I<16>, I<16>, I<16>>>;
         using Layouts    = std::tuple<
             std::tuple<col_major, row_major, row_major>>; //typename Base::TestLayoutsNT;
